fix(ci): Fixed GitHub action error reporting

+++ src/dialog/dialog_notification_settings.c
@@ -195,9 +195,9 @@ INT_PTR CALLBACK NotificationSettingsDlgProc(HWND hwndDlg, UINT msg, WPARAM wPar
                         !g_AppConfig.notification.display.disabled);
             
             int totalSeconds = g_AppConfig.notification.display.timeout_ms / 1000;
-            st.wHour = totalSeconds / 3600;
-            st.wMinute = (totalSeconds % 3600) / 60;
-            st.wSecond = totalSeconds % 60;
+            st.wHour = (WORD)(totalSeconds / 3600);
+            st.wMinute = (WORD)((totalSeconds % 3600) / 60);
+            st.wSecond = (WORD)(totalSeconds % 60);
             
             SendDlgItemMessage(hwndDlg, IDC_NOTIFICATION_TIME_EDIT, DTM_SETSYSTEMTIME, 
                               GDT_VALID, (LPARAM)&st);
